requirements.txt
# ML Platform Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# MLflow & Artifacts
mlflow>=2.14.0
boto3>=1.28.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
